+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
+### Added
+- add support for validating the `--link` flag in COPY instructions ([#103](https://github.com/rcjsuen/dockerfile-language-service/issues/103)
+
 ### Fixed
 - semantic tokens will no longer have negative values if quoted string content is inside a variable ([#104](https://github.com/rcjsuen/dockerfile-language-service/issues/104))